Some code cleanup, and several optimizations for speed, by exploiting some bit hacks. The changes are separated into different commits to make it a bit easier to trace them.

On my machine (Ubuntu, AMD Ryzen 7), the different optimization steps gave speed-ups as follows. The important metric is `enc/sec`, the number of encodings that are computed per second. Tests were run by encoding 100,000,000 kmers and their reverse complements (to check consistency), for a total of 200mio encodings per test. One with even `k`, one with odd `k`, as those behave differently.

```
initial pre my changes:
k==15, time: 8.899670s, enc/s: 22472744
k==16, time: 2.783863s, enc/s: 71842625

precomputed powers of 4
k==15, time: 2.792941s, enc/s: 71609099
k==16, time: 2.786097s, enc/s: 71785006

simplify encode prime
k==15, time: 2.767289s, enc/s: 72272894
k==16, time: 2.710502s, enc/s: 73787079

simplify encode single character in the middle
k==15, time: 2.719947s, enc/s: 73530844
k==16, time: 2.694091s, enc/s: 74236547

simplify encode specifying pair
k==15, time: 1.284619s, enc/s: 155688129
k==16, time: 1.063841s, enc/s: 187997971
```

Overall, the first big change is to pre-compute the `pow` calculation, as that only depends on `k`, but was conducted in every call, and is rather expensive. Then, some smaller improvements by optimizing the involved masks. Lastly, the second big change is the optimization of the specifying pair encoding, as that is by far the most common pattern.